module4/class6/requirements.txt
python-dotenv
langchain
langchain-openai
langgraph
pydantic
pandas
tabulate
langchain-qdrant
langchain_text_splitters
langchain-community
pymupdf
